refactor the useEvent
