--- web/src/hooks/useIsoMorphicEffect.ts
@@ -1,7 +1,0 @@
-import { useEffect, EffectCallback, DependencyList } from "react";
-
-const useIsoMorphicEffect = (effect: EffectCallback, deps?: DependencyList | undefined) => {
-  useEffect(effect, deps);
-};
-
-export default useIsoMorphicEffect;
